requirements.txt
uvicorn[standard]
fastapi
sqlalchemy
databases
aiosqlite
pyjwt
bcrypt